Per entry ignore of NS validation failures in CI testing
This adds the ability to ignore NS validation failures during CI
testing on a per-entry basis in the nses.yml files. If an entry
has the value `ignore_validation_errors: true`, then validation
errors will not cause CI failures.

This is primarily in order to be able to prevent CI failures
due to the persistent intermittent issues with validating
`dns-parking.com.`.